Reference Model: Conditionally build ERR_MFR register into the model
When IOPMP_MFR_EN=0, the ERR_MFR register should not be built into the
model and we replace it with new "reserved11". The record windows are
also not built into the model in this case.

The dedicated test case for MFR extension is also built only when
IOPMP_MFR_EN is set to 1.