Reference Model: Add iopmp_dev_t struct to hold states of an IOPMP device
Instead of using global variables to hold the state of the IOPMP, all of the
state has been added to a new struct called iopmp_dev_t. This is useful if
users want to model systems which have multiple IOPMPs because users can
just declare multiple iopmp_t instead of copying global states.

Now almost all APIs take an extra parameter which is a pointer to the
relevant iopmp_dev_t instance.