Reference Model: Refine the rule analyzer and matching logics
From the IOPMP spec, an entry qualifies as a matching entry if:
- Its region covers any byte of the transaction,
- It is associated with the RRID carried by the transaction; and
- It holds the highest priority among entries that meet the previous
  criteria.

Technically speaking, a "matching entry" doesn't mean that the entry
also grants transaction permission. The model needs to distinguish
between the "maching entry without granting permission" and "matching
entry with granting permission".

This commit splits standard error types from iopmpMatchStatus_t as
iopmpErrorType_t, and defines five status, including partial matching
and whether the matching entry grants transaction permission or not.

Now, iopmpMatchAddr() just check the address range of the entry.
Whether the entry "fully matches", "partially matches" or "doesn't
match" the bytes of the transaction, regardless of the entry is a
priority entry or not.

The iopmpCheckPerms() checks permission bits of the entry which "fully
matches" all bytes of the transaction, and simply returns whether the
entry "grants" or "doesn't grant" the transaction requested permission.

The final judgements are performed in iopmp_validate_access(). This
function gets the result from iopmpRuleAnalyzer(), determining if the
entry is non-priority entry or not, and judges if the transaction is
passed or the searching needs to be continued or stopped with an error
capture.